+++ kifi-backend/modules/scraper/app/com/keepit/scraper/actor/ScrapeProcessorActorImpl.scala
@@ -1,0 +1,87 @@
+package com.keepit.scraper.actor
+
+import akka.actor.{ ActorSystem, Props }
+import akka.pattern.ask
+import akka.util.Timeout
+import com.google.inject.{ Inject, Provider, Singleton }
+import com.keepit.common.concurrent.ExecutionContext
+import com.keepit.common.healthcheck.AirbrakeNotifier
+import com.keepit.common.logging.Logging
+import com.keepit.common.zookeeper.ServiceDiscovery
+import com.keepit.model.{ HttpProxy, NormalizedURI, PageInfo, ScrapeInfo }
+import com.keepit.scraper._
+import com.keepit.scraper.extractor._
+
+import scala.concurrent.Future
+import scala.concurrent.duration._
+import scala.util.{ Failure, Success }
+
+object ScraperMessages {
+  // Scrape: pure side-effects; Fetch: returns content (see fetchBasicArticle)
+  case class Fetch(url: String, proxyOpt: Option[HttpProxy], extractorProviderTypeOpt: Option[ExtractorProviderType])
+  case class Scrape(uri: NormalizedURI, info: ScrapeInfo, pageInfo: Option[PageInfo], proxyOpt: Option[HttpProxy]) {
+    override def toString = s"Scrape(uri=${uri.toShortString},info=${info.toShortString})"
+  }
+  case object QueueSize // informational; pulling
+}
+
+@Singleton
+class ScrapeProcessorActorImpl @Inject() (
+    airbrake: AirbrakeNotifier,
+    sysProvider: Provider[ActorSystem],
+    scrapeSupervisorProvider: Provider[ScrapeAgentSupervisor],
+    scrapeProcActorProvider: Provider[ScrapeAgent],
+    serviceDiscovery: ServiceDiscovery,
+    asyncHelper: ShoeboxDbCallbacks) extends ScrapeProcessor with Logging {
+
+  import ScraperMessages._
+
+  val PULL_THRESHOLD = Runtime.getRuntime.availableProcessors() / 2
+  val WARNING_THRESHOLD = 100
+
+  implicit val fj = ExecutionContext.fj
+  implicit val timeout = Timeout(15 seconds)
+
+  lazy val system = sysProvider.get
+  lazy val actor = system.actorOf(Props(scrapeSupervisorProvider.get), "scraper_supervisor")
+
+  def fetchBasicArticle(url: String, proxyOpt: Option[HttpProxy], extractorProviderTypeOpt: Option[ExtractorProviderType]): Future[Option[BasicArticle]] = {
+    actor.ask(Fetch(url, proxyOpt, extractorProviderTypeOpt))(Timeout(15 minutes)).mapTo[Option[BasicArticle]]
+  }
+
+  def asyncScrape(uri: NormalizedURI, info: ScrapeInfo, pageInfo: Option[PageInfo], proxyOpt: Option[HttpProxy]): Unit = {
+    actor ! Scrape(uri, info, pageInfo, proxyOpt)
+  }
+
+  private def getQueueSize(): Future[Int] = actor.ask(QueueSize).mapTo[Int]
+
+  override def pull(): Unit = {
+    getQueueSize() onComplete {
+      case Success(qSize) =>
+        if (qSize < PULL_THRESHOLD) {
+          log.info(s"[ScrapeProcessorActorImpl.pull] qSize=$qSize. Let's get some work.")
+          serviceDiscovery.thisInstance.map { inst =>
+            if (inst.isHealthy) {
+              asyncHelper.assignTasks(inst.id.id, PULL_THRESHOLD * 2).onComplete {
+                case Failure(t) =>
+                  log.error(s"[ScrapeProcessorActorImpl.pull(${inst.id.id})] Caught exception $t while pulling for tasks", t) // move along
+                case Success(requests) =>
+                  log.info(s"[ScrapeProcessorActorImpl.pull(${inst.id.id})] assigned (${requests.length}) scraping tasks: ${requests.map(r => s"[uriId=${r.uri.id},infoId=${r.scrapeInfo.id},url=${r.uri.url}]").mkString(",")} ")
+                  for (sr <- requests) {
+                    asyncScrape(sr.uri, sr.scrapeInfo, sr.pageInfoOpt, sr.proxyOpt)
+                  }
+              }(ExecutionContext.fj)
+            }
+          }
+        } else if (qSize > WARNING_THRESHOLD) {
+          airbrake.notify(s"qSize=${qSize} has exceeded threshold=$WARNING_THRESHOLD")
+        } else {
+          log.info(s"[ScrapeProcessorActorImpl.pull] qSize=${qSize}; Skip a round")
+        }
+      case Failure(e) =>
+        airbrake.notify(s"Failed to obtain qSize from supervisor; exception=$e; cause=${e.getCause}", e)
+    }
+  }
+
+}
+
